moved all styles from pages to styles
