+++ src/styles/coinDetail.js
