+++ src/styles/exchangeDetail.js
