+++ src/styles/nftDetail.js
