+++ src/styles/nfts.js
